requirements.txt
SQLAlchemy==1.2.14
numpy==1.19.2
nltk==3.4.5
hdbscan==0.8.26
germalemma==0.1.3
matplotlib==3.1.1
gensim==3.8.1
seaborn==0.9.0
plotly==4.11.0
pandas==1.1.0
cufflinks==0.17.3
dash==1.16.3
dash_core_components==1.12.1
dash_html_components==1.1.1
dash_table==4.10.1
pymysql==0.10.1
scikit_learn==0.23.2
umap-learn
pyLDAvis==2.1.2
sentence-transformers
MLflow==1.11.0
summa==1.2.0
psycopg2==2.8.6

